requirements.txt
discord.py>=2.3.2
httpx>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
